Backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-together
neo4j
chromadb
sentence-transformers
pydantic
python-dotenv